+++ flutter/lib/src/navigation/sentry_navigator_observer.dart
@@ -86,6 +86,7 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
         _routeNameExtractor = routeNameExtractor,
         _additionalInfoProvider = additionalInfoProvider,
         _native = SentryFlutter.native {
+    _isCreated = true;
     if (enableAutoTransactions) {
       _hub.options.sdk.addIntegration('UINavigationTracing');
     }
@@ -120,6 +121,11 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
   ISentrySpan? _transaction;
 
   static String? _currentRouteName;
+
+  static bool _isCreated = false;
+
+  @internal
+  static bool get isCreated => _isCreated;
 
   @internal
   static String? get currentRouteName => _currentRouteName;
@@ -224,7 +230,7 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
     }
 
     if (name == '/') {
-      name = 'root /';
+      name = rootScreenName;
     }
     final transactionContext = SentryTransactionContext(
       name,
@@ -366,6 +372,9 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
     _completedDisplayTracking = Completer();
     _timeToDisplayTracker?.clear();
   }
+
+  @internal
+  static const String rootScreenName = 'root /';
 }
 
 /// This class makes it easier to record breadcrumbs for events of Flutters