+++ flutter/lib/src/sentry_flutter.dart
@@ -34,10 +34,10 @@ typedef FlutterOptionsConfiguration = FutureOr<void> Function(
 mixin SentryFlutter {
   static const _channel = MethodChannel('sentry_flutter');
 
-  /// Represents the time when the main isolate is set up and ready to execute.
+  /// Represents the time when the Sentry init set up has started.
   @internal
   // ignore: invalid_use_of_internal_member
-  static DateTime mainIsolateStartTime = getUtcDateTime();
+  static DateTime? sentrySetupStartTime;
 
   static Future<void> init(
     FlutterOptionsConfiguration optionsConfiguration, {
@@ -47,6 +47,9 @@ mixin SentryFlutter {
     @internal RendererWrapper? rendererWrapper,
   }) async {
     final flutterOptions = SentryFlutterOptions();
+
+    // ignore: invalid_use_of_internal_member
+    sentrySetupStartTime ??= flutterOptions.clock();
 
     if (platformChecker != null) {
       flutterOptions.platformChecker = platformChecker;